% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbcsnowflake-package.R
\docType{package}
\name{adbcsnowflake-package}
\title{adbcsnowflake: Arrow Database Connectivity (ADBC) Snowflake Driver}
\description{
Provides a developer-facing interface to the Arrow Database Connectivity (ADBC) Snowflake driver.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/apache/arrow-adbc}
  \item Report bugs at \url{https://github.com/apache/arrow-adbc/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@dunnington.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

Authors:
\itemize{
  \item Apache Arrow \email{dev@arrow.apache.org} [copyright holder]
}

}
\keyword{internal}
